backend/data_fetching/requirements.txt
RPi.bme280==0.2.4
paho-mqtt==2.0.0
